# Backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Bootcamp 2025",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/activities/types": {
      "get": {
        "tags": ["Atividades"],
        "summary": "Listar tipos de atividades",
        "description": "Endpoint para listar os tipos de atividades disponíveis.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "3fa85764-5717-4562-b3fc-2c963f6a4fa6"
                      },
                      "name": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "image": {
                        "type": "string"
                      }
                    }
                  }
                },
                "example": [
                  {
                    "id": "3fa85764-5717-4562-b3fc-2c963f6a4fa6",
                    "name": "string",
                    "description": "string",
                    "image": "string"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Autenticação necessária."
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Esta conta foi desativada e não pode ser utilizada."
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Erro inesperado."
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/activities": {
      "get": {
        "tags": ["Atividades"],
        "summary": "Listar atividades com paginação, filtro por tipo e ordenação",
        "description": "Endpoint para listar de forma paginada as atividades registradas. Parâmetros opcionais: type (filtrar por um tipo específico de atividades), orderBy (campo pelo qual ordenar as atividades) e order (\"asc\" = crescente, \"desc\" = decrescente).",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Número da página.",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Tamanho da página.",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filtrar atividades por tipo.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Ordenar as atividades por um campo específico.",
            "schema": {
              "type": "string",
              "default": "createdAt"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Ordem pela qual ordenar as atividades.",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "page": {
                      "type": "integer",
                      "example": 0
                    },
                    "pageSize": {
                      "type": "integer",
                      "example": 0
                    },
                    "totalActivities": {
                      "type": "integer",
                      "example": 0
                    },
                    "totalPages": {
                      "type": "integer",
                      "example": 0
                    },
                    "previous": {
                      "type": "integer",
                      "example": 0
                    },
                    "next": {
                      "type": "integer",
                      "example": 0
                    },
                    "activities": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85764-5717-4562-b3fc-2c963f66afa6"
                          },
                          "title": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          },
                          "image": {
                            "type": "string"
                          },
                          "confirmationCode": {
                            "type": "string"
                          },
                          "participantCount": {
                            "type": "integer"
                          },
                          "address": {
                            "type": "object",
                            "properties": {
                              "latitude": {
                                "type": "number",
                                "format": "float"
                              },
                              "longitude": {
                                "type": "number",
                                "format": "float"
                              }
                            }
                          },
                          "scheduleDate": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "completedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "private": {
                            "type": "boolean"
                          },
                          "creator": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "name": {
                                "type": "string"
                              },
                              "avatar": {
                                "type": "string"
                              }
                            }
                          },
                          "userSubscriptionStatus": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "page": 0,
                  "pageSize": 0,
                  "totalActivities": 0,
                  "totalPages": 0,
                  "previous": 0,
                  "next": 0,
                  "activities": [
                    {
                      "id": "3fa85764-5717-4562-b3fc-2c963f66afa6",
                      "title": "string",
                      "description": "string",
                      "type": "string",
                      "image": "string",
                      "confirmationCode": "string",
                      "participantCount": 0,
                      "address": {
                        "latitude": 0,
                        "longitude": 0
                      },
                      "scheduleDate": "2025-03-20T18:00:01.4032",
                      "createdAt": "2025-03-20T18:00:01.4032",
                      "completedAt": "2025-03-20T18:00:01.4032",
                      "private": true,
                      "creator": {
                        "id": "3fa85f64-5717-4562-b3fc-2c963f6a4fa6",
                        "name": "string",
                        "avatar": "string"
                      },
                      "userSubscriptionStatus": "string"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Autenticação necessária."
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Esta conta foi desativada e não pode ser utilizada."
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Erro inesperado."
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/activity/all": {
      "get": {
        "tags": ["Atividades"],
        "summary": "Listar todas as atividades com filtro por tipo e ordenação",
        "description": "Endpoint para listar todas as atividades registradas. Parâmetros opcionais: type (filtrar por um tipo específico de atividades), orderBy (campo pelo qual ordenar as atividades) e order (\"asc\" = crescente, \"desc\" = decrescente).",
        "parameters": [
          {
            "name": "typeId",
            "in": "query",
            "description": "Filtrar atividades por tipo.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Ordenar as atividades por um campo específico.",
            "schema": {
              "type": "string",
              "default": "createdAt"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Ordem pela qual ordenar as atividades.",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                      },
                      "title": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string"
                      },
                      "image": {
                        "type": "string"
                      },
                      "confirmationCode": {
                        "type": "string"
                      },
                      "participantCount": {
                        "type": "integer"
                      },
                      "address": {
                        "type": "object",
                        "properties": {
                          "latitude": {
                            "type": "number",
                            "format": "float"
                          },
                          "longitude": {
                            "type": "number",
                            "format": "float"
                          }
                        }
                      },
                      "scheduleDate": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "completedAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "private": {
                        "type": "boolean"
                      },
                      "creator": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "name": {
                            "type": "string"
                          },
                          "avatar": {
                            "type": "string"
                          }
                        }
                      },
                      "userSubscriptionStatus": {
                        "type": "string"
                      }
                    }
                  }
                },
                "example": [
                  {
                    "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                    "title": "string",
                    "description": "string",
                    "type": "string",
                    "image": "string",
                    "confirmationCode": "string",
                    "participantCount": 0,
                    "address": {
                      "latitude": 0,
                      "longitude": 0
                    },
                    "scheduleDate": "2025-03-20T18:04:07.117Z",
                    "createdAt": "2025-03-20T18:04:07.117Z",
                    "completedAt": "2025-03-20T18:04:07.117Z",
                    "private": true,
                    "creator": {
                      "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                      "name": "string",
                      "avatar": "string"
                    },
                    "userSubscriptionStatus": "string"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Autenticação necessária."
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Esta conta foi desativada e não pode ser utilizada."
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Erro inesperado."
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/activities/{id}/conclude": {
      "put": {
        "tags": ["Atividades"],
        "summary": "Concluir uma atividade",
        "description": "Endpoint para concluir uma atividade.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da atividade a ser concluída",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "message": "Atividade concluída com sucesso."
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Autenticação necessária."
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Esta conta foi desativada e não pode ser utilizada."
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Atividade não encontrada."
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Erro inesperado."
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}